+++ packages/frontend/src/components/composer/Composer.tsx
@@ -405,6 +405,7 @@ const Composer = forwardRef<
               enterKeySends={settingsStore?.desktopSettings.enterKeySends}
               sendMessage={composerSendMessage}
               chatId={chatId}
+              chatName={selectedChat.name}
               updateDraftText={updateDraftText}
               onPaste={handlePaste}
             />