+++ packages/frontend/src/components/chat/ChatListItemRow.tsx
@@ -57,7 +57,16 @@ export const ChatListItemRowChat = React.memo<{
   const chatId = chatListIds[index]
 
   return (
-    <li style={style}>
+    <li
+      style={style}
+      // TODO looks like screen-readers (at least NVDA) do not announce
+      // this "position" thing. Probably because the `<li>` itself
+      // is not focusable.
+      // Should we apply these attributes to the button?
+      // And get rid of the wrapper `<li>` entirely?
+      aria-posinset={index + 1}
+      aria-setsize={chatListIds.length}
+    >
       <ChatListItem
         isSelected={selectedChatId === chatId}
         chatListItem={chatCache[chatId] || undefined}
@@ -98,7 +107,11 @@ export const ChatListItemRowContact = React.memo<{
       }}
     />
   ) : (
-    <li style={style}>
+    <li
+      style={style}
+      aria-posinset={index + 1}
+      aria-setsize={contactIds.length}
+    >
       <PlaceholderChatListItem />
     </li>
   )
@@ -116,7 +129,11 @@ export const ChatListItemRowMessage = React.memo<{
   const accountId = selectedAccountId()
 
   return (
-    <li style={style}>
+    <li
+      style={style}
+      aria-posinset={index + 1}
+      aria-setsize={messageResultIds.length}
+    >
       {messageSearchResult ? (
         <ChatListItemMessageResult
           queryStr={queryStr || ''}