+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 
 - keep the order of contacts when calling getContactsByIds #4651
 
+### Fixed
+- accessibility: don't announce "padlock" on messages
+
 <a id="1_54_0"></a>
 
 ## [1.54.0] - 2025-02-15