+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 - fix unread count on "jump down" button not clearing when all messages are read #4648
 - keep the order of contacts when calling getContactsByIds #4651
 
+### Fixed
+- accessibility: add `aria-label`s to unread counters
+
 <a id="1_54_0"></a>
 
 ## [1.54.0] - 2025-02-15