+++ CHANGELOG.md
@@ -6,11 +6,9 @@
 
 ### Changed
 
-- keep the order of contacts when calling getContactsByIds #4651, #4652
-
 ### Fixed
 - fix unread count on "jump down" button not clearing when all messages are read #4648
-- keep the order of contacts when calling getContactsByIds #4651
+- keep the order of contacts when calling getContactsByIds #4651, #4652
 - improve accessibility #4655, #4656, #4661, #4662
 
 <a id="1_54_0"></a>